いらないシーン、スクリプトの削除 UI表示のコメントアウト
