--- Assets/07鉄瓶ゴシック.otf.meta
@@ -1,22 +1,0 @@
-fileFormatVersion: 2
-guid: 6fc4fc251c697f24dbe5145eae118913
-TrueTypeFontImporter:
-  externalObjects: {}
-  serializedVersion: 4
-  fontSize: 16
-  forceTextureCase: -2
-  characterSpacing: 0
-  characterPadding: 1
-  includeFontData: 1
-  fontName: 07TetsubinGothic
-  fontNames:
-  - 07TetsubinGothic
-  fallbackFontReferences: []
-  customCharacters: 
-  fontRenderingMode: 0
-  ascentCalculationMode: 1
-  useLegacyBoundsCalculation: 0
-  shouldRoundAdvanceValue: 1
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
