--- Assets/Assets/Materials/ketumaro.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 56fc5ccd9f7c93d41bcf0b572a3de045
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
