--- Assets/Assets/Prefabs/FakeFigure.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 0e53eaef544d2da48a44384ddba4cb28
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
