--- Assets/Assets/Prefabs/Node2D.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 46638ef8ea12c2d4c8d8e24f8a47c5d6
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
