--- Assets/Assets/Scenes/BoardScene2D.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 6baec6ffff48adb4faffcf5dd7e0750b
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
