--- Assets/Assets/Scenes/DiskEditScene.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: ddbe2ccfa127ecc4ab93a6beddab2850
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
