--- Assets/Assets/Scenes/HALTest.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 23c6c1d0397fb2e4ea5aaaef7ab8b41f
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
