--- Assets/Assets/Scenes/SpinScene2D.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 34eed99dfb865414285f004eaec9073e
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
