--- Assets/Assets/Scenes/TitleScene.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: ce5f292aee0feb943932fb493c5c42a3
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
