+++ Assets/Assets/Scripts/BoardMaster/BoardController.cs
@@ -4,43 +4,54 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
 using UnityEngine.SceneManagement;
+using TMPro;
 
 public class BoardController : MonoBehaviour
 {
     //あるノードから別のノードへのエッジをListで表現
     //edgesは配列であり、各成分がint型のList
     private List<int>[] edges = new List<int>[44];
-    private int fieldNodeNumber = 28;
-    private int benchNodeNumber = 12;
-    private int pcNodeNumber = 4;
+    private int[][] entryNodeID = new int[2][];
+    private int[] goalNodeID = new int[2];
+    private int[][] benchNodeID = new int[2][];
+    private int[][] pcNodeID = new int[2][];
+
     //第一要素がプレイヤー番号(0 or 1)、第二要素がfigureIDOnBoard
     //ゲーム開始時に敵味方それぞれのフィギュアを認識してfigureIDOnBoardを振る
     private List<GameObject>[] figures = new List<GameObject>[2];
+
     //ノードの親要素
     [SerializeField] private GameObject nodes;
 
-    //LIneRenderer用のやつ
+    //エッジ描画のためのprefab
     [SerializeField] private GameObject drawEdgePrefab;
 
-
+    //UI
+    [SerializeField] private GameObject startText;
+    [SerializeField] private List<GameObject> playerTurnText;
+    [SerializeField] private GameObject restTurnText;
+    [SerializeField] private GameObject gameEndText;
+    [SerializeField] private GameObject turnEndButton;
     //そのノードにつく前はどこにいたのかを表す
     private int[] prevNode = new int[44];
 
     //移動、バトルの主体となるフィギュア（"使用"するフィギュア）
     private GameObject currentFigure = null;
+    //歩行範囲と攻撃範囲
     private List<int> walkCandidates = new List<int>();
     private List<int> attackCandidates = new List<int>();
     //バトル相手のフィギュア
-    private GameObject opponentFigure;
+    private GameObject opponentFigure = null;
     //効果の対象など
-    private GameObject targetFigure;
+    private GameObject targetFigure = null;
 
     //どちらのターンかを表す。{0,1}で定められる
-    private int turnNumber;
-
-    private int restTurn;
+    private int turnNumber = 0;
+    //残りターン数
+    private int restTurn = 300;
     public enum PhaseState
     {
+        TurnStart,
         Normal,
         FigureSelected,
         Walking,
@@ -48,29 +59,41 @@ public class BoardController : MonoBehaviour
         ConfirmFigure,
         Battle,
         AfterBattle,
-        End
+        TurnEnd,
+        GameEnd
     };
     //ゲームのの状態変数
     private PhaseState phaseState;
     // Start is called before the first frame update
     void Start()
     {
+
         //ボードのデータ構造作成（ゲームの最初に1回だけ呼ばれる
         CreateBoard();
         EdgeDraw();
         DivideFigures();
-        SetPhaseState(PhaseState.Normal);
+        /* UIカット
+        yield return FadeInOut(startText, 1);
+        Destroy(startText);
+        */
         //暫定的にプレイヤー0のターンに固定
         turnNumber = 0;
-        //実際は300て表示してから減らすが
-        restTurn = 299;
-        Debug.Log("プレイヤー" + turnNumber + "のターンです");
+        StartCoroutine(SetPhaseState(PhaseState.TurnStart));
+
+
+        //Debug.Log("プレイヤー" + turnNumber + "のターンです");
         //Debug.Log(figures.Length);
+
+        // イベントにイベントハンドラーを追加
+        SceneManager.sceneLoaded += SceneLoaded;
     }
     void Update()
     {
 
     }
+
+    //ボード、フィギュアの初期化処理
+    //
     //ボードのデータ構造を表現
     void CreateBoard()
     {
@@ -122,38 +145,49 @@ public class BoardController : MonoBehaviour
         edges[36].Add(0); edges[36].Add(6);
         edges[37].Add(0); edges[37].Add(6);
         edges[38].Add(0); edges[38].Add(6);
-        edges[38].Add(0); edges[38].Add(6);
+        edges[39].Add(0); edges[39].Add(6);
 
-        //PC
+        //第一成分のインデックスはプレイヤーID、第二成分のインデックス0番が左エントリー, インデックス1番が右エントリー
+        for (int i = 0; i < entryNodeID.Length; i++) entryNodeID[i] = new int[2];
+        entryNodeID[0][0] = 21; entryNodeID[0][1] = 27;
+        entryNodeID[1][0] = 6; entryNodeID[1][1] = 0;
+
+        //第一成分のインデックスはプレイヤーID
+        goalNodeID[0] = 24; goalNodeID[1] = 3;
+
+        //第一成分のインデックスはプレイヤーID, 各要素figureParameterのbenchIDに対応
+        for (int i = 0; i < benchNodeID.Length; i++) benchNodeID[i] = new int[6];
+        for (int i = 0; i < 6; i++) benchNodeID[0][i] = 28 + i;
+        for (int i = 0; i < 6; i++) benchNodeID[1][i] = 28 + 6 + i;
+
+        //第一成分のインデックスはプレイヤーID , 第二成分のインデックス0番は先に入るPC, インデックス1番は後に入るPC
+        for (int i = 0; i < pcNodeID.Length; i++) pcNodeID[i] = new int[2];
+        pcNodeID[0][0] = 40; pcNodeID[0][1] = 41;
+        pcNodeID[1][0] = 42; pcNodeID[1][1] = 43;
     }
     //エッジの描画
     void EdgeDraw()
     {
-        for(int i = 0; i < fieldNodeNumber; i++)
+        for (int i = 0; i < 28; i++)
         {
             for (int j = 0; j < edges[i].Count; j++)
             {
-                float offset = 0.01f;
                 GameObject obj = Instantiate(drawEdgePrefab, transform.position, Quaternion.identity);
                 obj.transform.SetParent(transform);
                 LineRenderer line = drawEdgePrefab.GetComponent<LineRenderer>();
+                line.sortingOrder = 1;
                 line.startWidth = 0.05f;
                 line.endWidth = 0.05f;
 
                 //頂点の数を決める
                 line.positionCount = 2;
-                
-                Vector3 startDrawPosition = new Vector3(nodes.transform.GetChild(i).position.x,
-                                                        nodes.transform.GetChild(i).position.y - offset,
-                                                        nodes.transform.GetChild(i).position.z);
 
-                Vector3 endDrawPosition = new Vector3(nodes.transform.GetChild(edges[i][j]).position.x,
-                                                      nodes.transform.GetChild(edges[i][j]).position.y - offset,
-                                                      nodes.transform.GetChild(edges[i][j]).position.z);
+                Vector3 startDrawPosition = nodes.transform.GetChild(i).position;
+
+                Vector3 endDrawPosition = nodes.transform.GetChild(edges[i][j]).position;
                 line.SetPosition(0, startDrawPosition);
                 line.SetPosition(1, endDrawPosition);
-                
-                
+
             }
         }
     }
@@ -172,7 +206,7 @@ public class BoardController : MonoBehaviour
             {
                 figureParameter.SetFigureIDOnBoard(figures[0].Count);
                 figures[0].Add(obj);
-                obj.GetComponent<Renderer>().material.color = Color.blue;
+                obj.transform.Find("FigureBack2").GetComponent<SpriteRenderer>().color = Color.blue;
                 //Debug.Log(figures[0][figureParameter.GetFigureIDOnBoard()]);
             }
             else if (figureParameter.GetPlayerID() == 1)
@@ -180,30 +214,34 @@ public class BoardController : MonoBehaviour
                 figureParameter.SetFigureIDOnBoard(figures[1].Count);
                 figures[1].Add(obj);
                 obj.transform.Rotate(0, 0, 180f);
-                obj.GetComponent<Renderer>().material.color = Color.red;
+                obj.transform.Find("FigureBack2").GetComponent<SpriteRenderer>().color = Color.red;
             }
-            for(int i = 0; i < figures[0].Count; i++)
+            for (int i = 0; i < figures[0].Count; i++)
             {
-                figures[0][i].GetComponent<FigureParameter>().SetBenchID(fieldNodeNumber + i);
-                figures[0][i].GetComponent<FigureParameter>().SetPosition(fieldNodeNumber + i);
+                figures[0][i].GetComponent<FigureParameter>().SetBenchID(28 + i);
+                figures[0][i].GetComponent<FigureParameter>().SetPosition(28 + i);
             }
             for (int i = 0; i < figures[1].Count; i++)
             {
-                figures[1][i].GetComponent<FigureParameter>().SetBenchID(fieldNodeNumber + 6 + i );
-                figures[1][i].GetComponent<FigureParameter>().SetPosition(fieldNodeNumber + 6 + i);
+                figures[1][i].GetComponent<FigureParameter>().SetBenchID(28 + 6 + i);
+                figures[1][i].GetComponent<FigureParameter>().SetPosition(28 + 6 + i);
             }
         }
     }
+    //
+    //
 
+
+    //
+    //
     //マスク配列の作成。
     //とりあえず場にいるすべてのフィギュアを障害物と認識するだけだが、すり抜けとかの実装の際要拡張
-    
     bool[] CreateNormalMask()
     {
         //該当するノードのmaskNodeがtrueの時すり抜け可能、falseなら障害物扱い
-        bool[] maskNode = new bool[fieldNodeNumber + benchNodeNumber + pcNodeNumber];
+        bool[] maskNode = new bool[44];
         for (int i = 0; i < maskNode.Length; i++) maskNode[i] = true;
-        for(int i = 0; i < figures.Length; i++)
+        for (int i = 0; i < figures.Length; i++)
         {
             for (int j = 0; j < figures[i].Count; j++)
             {
@@ -218,19 +256,18 @@ public class BoardController : MonoBehaviour
     }
 
     //始点から各ノードへの距離を計算
-
     (int[] distances, int[] prevNode) CalculateDistance(int _startNode, bool[] _maskNode)
     {
-        int[] distances = new int[fieldNodeNumber + benchNodeNumber + pcNodeNumber];
-        int[] prevNode = new int[fieldNodeNumber + benchNodeNumber + pcNodeNumber];
+        int[] distances = new int[44];
+        int[] prevNode = new int[44];
         //幅優先探索を素直に実装
         Queue<int> q = new Queue<int>();
 
         //各ノードの到着フラグ
-        bool[] reachedFlag = new bool[fieldNodeNumber + benchNodeNumber + pcNodeNumber];
+        bool[] reachedFlag = new bool[44];
 
         //distancesと到着フラグの初期化
-        for (int i = 0; i < fieldNodeNumber + benchNodeNumber + pcNodeNumber; i++)
+        for (int i = 0; i < 44; i++)
         {
             distances[i] = 99;
             reachedFlag[i] = false;
@@ -272,7 +309,7 @@ public class BoardController : MonoBehaviour
     List<int> FindCandidateofDestinaitonEqual(int _moveNumber, int[] _distances)
     {
         List<int> candidates = new List<int>();
-        for (int i = 0; i < fieldNodeNumber; i++)
+        for (int i = 0; i < 28; i++)
         {
             //始点から各ノードへの距離がmoveNumberと等しいであるノードを全て格納
             if (_distances[i] == _moveNumber) candidates.Add(i);
@@ -283,14 +320,14 @@ public class BoardController : MonoBehaviour
     }
 
     //始点からの距離がmoveNumber以下のノードを出力
-    List<int> FindCandidateofDestinaitonLessThan(int _moveNumber, int[]_distances)
+    List<int> FindCandidateofDestinaitonLessThan(int _moveNumber, int[] _distances)
     {
         List<int> candidates = new List<int>();
-        for (int i = 0; i < fieldNodeNumber; i++)
+        for (int i = 0; i < 28; i++)
         {
             //始点から各ノードへの距離がmoveNumber以下であるノードを全て格納
             if (_distances[i] <= _moveNumber) candidates.Add(i);
-        }       
+        }
         return candidates;
     }
 
@@ -331,25 +368,24 @@ public class BoardController : MonoBehaviour
 
             //状態変数更新
             //moveCandidates, walkCandidates, prevNodeの更新
-            SetPhaseState(PhaseState.FigureSelected);
+            StartCoroutine(SetPhaseState(PhaseState.FigureSelected));
         }
 
         else if (phaseState == PhaseState.FigureSelected)
         {
             //遷移2番: FigureSelect → Normal
             //遷移条件: クリックしたフィギュアがcurrentFigureと同じ
-            if(figures[_playerID][_figureIDOnBoard] == currentFigure)
+            if (figures[_playerID][_figureIDOnBoard] == currentFigure)
             {
-               　SetPhaseState(PhaseState.Normal);
+                StartCoroutine(SetPhaseState(PhaseState.Normal));
             }
             //遷移3番: FigureSelect→FigureSelect
-            else if(figures[_playerID][_figureIDOnBoard].GetComponent<FigureParameter>().GetPlayerID() == turnNumber)
+            else if (figures[_playerID][_figureIDOnBoard].GetComponent<FigureParameter>().GetPlayerID() == turnNumber)
             {
                 //ノードの色を初期化
-                for (int i = 0; i < fieldNodeNumber; i++)
+                for (int i = 0; i < 28; i++)
                 {
-                    Image image = nodes.transform.GetChild(i).GetComponent<Image>();
-                    image.color = Color.white;
+                    nodes.transform.GetChild(i).GetComponent<SpriteRenderer>().color = Color.white;
                 }
                 //光沢を初期化
                 currentFigure.GetComponent<Renderer>().material.SetFloat("_Metallic", 0.0f);
@@ -357,10 +393,10 @@ public class BoardController : MonoBehaviour
                 //currentFigure更新
                 currentFigure = figures[_playerID][_figureIDOnBoard];
                 //状態変数更新
-                SetPhaseState(PhaseState.FigureSelected);
+                StartCoroutine(SetPhaseState(PhaseState.FigureSelected));
             }
             //相手のコマをタッチしたとき
-            else if(figures[_playerID][_figureIDOnBoard].GetComponent<FigureParameter>().GetPlayerID() != turnNumber)
+            else if (figures[_playerID][_figureIDOnBoard].GetComponent<FigureParameter>().GetPlayerID() != turnNumber)
             {
                 bool attackOK = false;
                 foreach (int i in attackCandidates)
@@ -371,13 +407,12 @@ public class BoardController : MonoBehaviour
                     }
                 }
                 //敵が攻撃範囲内にいなかったとき、遷移3番: FigureSelect→FigureSelect
-                if(attackOK == false)
+                if (attackOK == false)
                 {
                     //ノードの色を初期化
-                    for (int i = 0; i < fieldNodeNumber; i++)
+                    for (int i = 0; i < 28; i++)
                     {
-                        Image image = nodes.transform.GetChild(i).GetComponent<Image>();
-                        image.color = Color.white;
+                        nodes.transform.GetChild(i).GetComponent<SpriteRenderer>().color = Color.white;
                     }
                     //光沢を初期化
                     currentFigure.GetComponent<Renderer>().material.SetFloat("_Metallic", 0.0f);
@@ -386,13 +421,13 @@ public class BoardController : MonoBehaviour
                     //currentFigure更新
                     currentFigure = figures[_playerID][_figureIDOnBoard];
                     //状態変数更新
-                    SetPhaseState(PhaseState.FigureSelected);
+                    StartCoroutine(SetPhaseState(PhaseState.FigureSelected));
                 }
                 //敵が攻撃範囲にいたとき、遷移5番: FigureSelect→Batle
                 else
                 {
                     opponentFigure = figures[_playerID][_figureIDOnBoard];
-                    SetPhaseState(PhaseState.Battle);
+                    StartCoroutine(SetPhaseState(PhaseState.Battle));
                 }
             }
         }
@@ -410,7 +445,7 @@ public class BoardController : MonoBehaviour
             //遷移7番: AfterWalk→ConfirmFigure
             else if (figures[_playerID][_figureIDOnBoard].GetComponent<FigureParameter>().GetPlayerID() == turnNumber)
             {
-                SetPhaseState(PhaseState.ConfirmFigure);
+                StartCoroutine(SetPhaseState(PhaseState.ConfirmFigure));
             }
 
             else if (figures[_playerID][_figureIDOnBoard].GetComponent<FigureParameter>().GetPlayerID() != turnNumber)
@@ -427,19 +462,19 @@ public class BoardController : MonoBehaviour
                 if (attackOK == false)
                 {
                     //状態変数更新
-                    SetPhaseState(PhaseState.ConfirmFigure);
+                    StartCoroutine(SetPhaseState(PhaseState.ConfirmFigure));
                 }
                 //敵が攻撃範囲にいたとき、遷移11番: FigureSelect→Batle
                 else
                 {
                     opponentFigure = figures[_playerID][_figureIDOnBoard];
-                    SetPhaseState(PhaseState.Battle);
+                    StartCoroutine(SetPhaseState(PhaseState.Battle));
                 }
             }
 
         }
-        
-        else if(phaseState == PhaseState.ConfirmFigure)
+
+        else if (phaseState == PhaseState.ConfirmFigure)
         {
             //遷移8番: ConfirmFigure →ConfirmFigure
             if (figures[_playerID][_figureIDOnBoard] != currentFigure)
@@ -449,11 +484,12 @@ public class BoardController : MonoBehaviour
             //遷移9番: ConfirmFigure→AfterWalk
             else if (figures[_playerID][_figureIDOnBoard] == currentFigure)
             {
-                SetPhaseState(PhaseState.AfterWalk);
+                StartCoroutine(SetPhaseState(PhaseState.AfterWalk));
             }
         }
+        else return;
     }
-    
+
     //遷移4番 FigureSelect → Walking
     //FigureSelect状態でcandidates内のノードがクリックされたら、選択中のフィギュアに経路情報を渡す
     public IEnumerator NodeClicked(int _nodeID)
@@ -465,32 +501,42 @@ public class BoardController : MonoBehaviour
                 {
                     if (_nodeID == walkCandidates[i])
                     {
+                        StartCoroutine(SetPhaseState(PhaseState.Walking));
                         //ここの引数のprevNodeはFigureSelectedが呼ばれたときに格納されているよ
- 
+
                         Stack<int> route = DecideRoute(_nodeID, prevNode);
-                       //コルーチンを使った移動に変えた
-                        if (currentFigure.GetComponent<FigureParameter>().GetPosition() >= fieldNodeNumber)
+                        //コルーチンを使った移動に変えた
+                        if (currentFigure.GetComponent<FigureParameter>().GetPosition() >= 28)
                         {
                             route.Pop();
                             yield return StartCoroutine(currentFigure.GetComponent<FigureController>().FigureOneStepWalk(route.Peek()));
                         }
-                        if(route.Count >= 2)
+                        if (route.Count >= 2)
                         {
                             yield return StartCoroutine(currentFigure.GetComponent<FigureController>().Figurewalk(route));
                         }
-                        SetPhaseState(BoardController.PhaseState.AfterWalk);
+                        StartCoroutine(SetPhaseState(PhaseState.AfterWalk));
                     }
                 }
         }
     }
-    
+
     //ゲームの状態変数のゲッター、セッター
-    public void SetPhaseState(PhaseState _tempState)
+    public IEnumerator SetPhaseState(PhaseState _tempState)
     {
         phaseState = _tempState;
         Debug.Log(GetPhaseState());
 
-        if (phaseState == PhaseState.Normal)
+        //ターンの開始時
+        if(phaseState == PhaseState.TurnStart)
+        {
+            /*UIカット
+            restTurnText.GetComponent<TextMeshProUGUI>().text = restTurn.ToString();
+            yield return FadeInOut(playerTurnText[turnNumber], 0.5f);
+            */
+            StartCoroutine(SetPhaseState(PhaseState.Normal));
+        }
+        else if (phaseState == PhaseState.Normal)
         {
             //各種変数の初期化
             //光沢を消す
@@ -504,13 +550,12 @@ public class BoardController : MonoBehaviour
             walkCandidates.Clear();
             attackCandidates.Clear();
             //ノードの色を初期化
-            for (int i = 0; i < fieldNodeNumber; i++)
+            for (int i = 0; i < 28; i++)
             {
-                Image image = nodes.transform.GetChild(i).GetComponent<Image>();
-                image.color = Color.white;
+                nodes.transform.GetChild(i).GetComponent<SpriteRenderer>().color = Color.white;
             }
         }
-        else if(phaseState == PhaseState.FigureSelected)
+        else if (phaseState == PhaseState.FigureSelected)
         {
             //選択したフィギュアを目立たせる
             currentFigure.GetComponent<Renderer>().material.SetFloat("_Metallic", 1.0f);
@@ -525,30 +570,28 @@ public class BoardController : MonoBehaviour
             //この計算の時にDecideRouteに必要なprevNodeを記録
             prevNode = calculateWalkDistance.prevNode;
             //現在地からmp以内で移動可能な全ノードの色を紫色にする
-            if(restTurn == 299)
+            if (restTurn == 300)
             {
                 walkCandidates = FindCandidateofDestinaitonLessThan(currentFigureParameter.GetMp() - 1, calculateWalkDistance.distances);
             }
             else walkCandidates = FindCandidateofDestinaitonLessThan(currentFigureParameter.GetMp(), calculateWalkDistance.distances);
             for (int i = 0; i < walkCandidates.Count; i++)
             {
-                Image image = nodes.transform.GetChild(walkCandidates[i]).GetComponent<Image>();
-                image.color = Color.magenta;
+                nodes.transform.GetChild(walkCandidates[i]).GetComponent<SpriteRenderer>().color = Color.magenta;
             }
 
             //攻撃範囲の計算
-            bool[] maskAttackNode = new bool[fieldNodeNumber + benchNodeNumber + pcNodeNumber];
+            bool[] maskAttackNode = new bool[44];
             for (int i = 0; i < maskAttackNode.Length; i++) maskAttackNode[i] = true;
             var calculateAttackDistance = CalculateDistance(currentFigureParameter.GetPosition(), maskAttackNode);
             attackCandidates = FindCandidateofDestinaitonLessThan(currentFigureParameter.GetAttackRange(), calculateAttackDistance.distances);
         }
-        else if(phaseState == PhaseState.AfterWalk)
+        else if (phaseState == PhaseState.AfterWalk)
         {
             //ノードの色を初期化
-            for (int i = 0; i < fieldNodeNumber; i++)
+            for (int i = 0; i < 28; i++)
             {
-                Image image = nodes.transform.GetChild(i).GetComponent<Image>();
-                image.color = Color.white;
+                nodes.transform.GetChild(i).GetComponent<SpriteRenderer>().color = Color.white;
             }
 
             //周囲に敵が誰もいなければターンエンド
@@ -560,7 +603,7 @@ public class BoardController : MonoBehaviour
             if (currentFigureParameter.GetPlayerID() == 0) opponentID = 1;
             else opponentID = 0;
 
-            bool[] maskAttackNode = new bool[fieldNodeNumber + benchNodeNumber + pcNodeNumber];
+            bool[] maskAttackNode = new bool[28];
             for (int i = 0; i < maskAttackNode.Length; i++) maskAttackNode[i] = true;
             var calculateAttackDistance = CalculateDistance(currentFigureParameter.GetPosition(), maskAttackNode);
             attackCandidates = FindCandidateofDestinaitonLessThan(currentFigureParameter.GetAttackRange(), calculateAttackDistance.distances);
@@ -568,47 +611,75 @@ public class BoardController : MonoBehaviour
 
             for (int i = 0; i < figures[opponentID].Count; i++)
             {
-                foreach(int j in attackCandidates)
+                foreach (int j in attackCandidates)
                 {
-                    if(j == figures[opponentID][i].GetComponent<FigureParameter>().GetPosition())
+                    if (j == figures[opponentID][i].GetComponent<FigureParameter>().GetPosition())
                     {
-
                         opponentExistInAttackCandidates = true;
-                        //Debug.Log(opponentExistInAttackCandidates);
                     }
+                    
                 }
             }
-            if (opponentExistInAttackCandidates == false) SetPhaseState(PhaseState.End);
+            //今ゴール処理は通常移動後しか考えてない。ワザの効果で移動するときもな
+            if (currentFigure.GetComponent<FigureParameter>().GetPosition() == goalNodeID[opponentID])
+            {
+                StartCoroutine(SetPhaseState(PhaseState.GameEnd));
+            }
+            else if (opponentExistInAttackCandidates == false) StartCoroutine(SetPhaseState(PhaseState.TurnEnd));
+            
+            else if (opponentExistInAttackCandidates == true)
+            {
+                turnEndButton.SetActive(true);
+            }
 
         }
-        else if(phaseState == PhaseState.Battle)
+        else if (phaseState == PhaseState.Battle)
         {
-            //ノードの色を初期化
-            for (int i = 0; i < fieldNodeNumber; i++)
-            {
-                Image image = nodes.transform.GetChild(i).GetComponent<Image>();
-                image.color = Color.white;
-            }
-            //シーンを移動させる処理をここに
             SceneManager.LoadScene("SpinScene");
-            //一時的にここですぐにAfterBattleへ
-            //SetPhaseState(PhaseState.AfterBattle);
         }
-        else if(phaseState == PhaseState.AfterBattle)
+        else if (phaseState == PhaseState.AfterBattle)
         {
             //バトル後の処理をここに
-            //一時的にここですぐにEndへ
-            SetPhaseState(PhaseState.End);
+
+            //バトル結果の受け取り
+            var BattleResult = SpinController.GetBattleResult();
+            int result = BattleResult.Item1;
+            bool currentMoveAwake = BattleResult.Item2;
+            bool opponentMoveAwake = BattleResult.Item3;
+            bool currentDeath = BattleResult.Item4;
+            bool opponentDeath = BattleResult.Item5;
+            Debug.Log("boardで" + BattleResult);
+            //if(currentMoveAwake){}
+            //if(opponentMoveAwake){}
+            if (currentDeath)
+            {
+                yield return Death(currentFigure);
+            }
+            if (opponentDeath)
+            {
+                yield return Death(opponentFigure);
+            }
+            StartCoroutine(SetPhaseState(PhaseState.TurnEnd));
         }
-        else if(phaseState == PhaseState.End)
+        else if (phaseState == PhaseState.TurnEnd)
         {
             //相手のターンにして残りのターン数を更新
             if (turnNumber == 0) turnNumber = 1;
             else if (turnNumber == 1) turnNumber = 0;
             restTurn--;
-            Debug.Log("プレイヤー" + turnNumber + "のターンです");
-            SetPhaseState(PhaseState.Normal);
+            turnEndButton.SetActive(false);
+            //Debug.Log("プレイヤー" + turnNumber + "のターンです");
+            StartCoroutine(SetPhaseState(PhaseState.TurnStart));
         }
+        else if(phaseState == PhaseState.GameEnd)
+        {
+            //もう少し真面目にかけ
+            //実際はどっかのタイミングで全部のフィギュアの位置調べて相手のゴールにいればなんちゃらみたいな感じかなあ知らんけど
+                gameEndText.SetActive(true);
+                gameEndText.GetComponent<TextMeshProUGUI>().text = "PLAYER" + currentFigure.GetComponent<FigureParameter>().GetPlayerID() + " WIN!";
+        }
+        else yield return null;
+        //yield return null;
     }
 
     public PhaseState GetPhaseState()
@@ -624,5 +695,134 @@ public class BoardController : MonoBehaviour
     {
         return opponentFigure;
     }
-    // Update is called once per frame
+
+    public IEnumerator Death(GameObject _figure)
+    {
+        GameObject backBenchFigure = null;
+        GameObject moveAnotherPCFigure = null;
+        int playerID = _figure.GetComponent<FigureParameter>().GetPlayerID();
+
+        for (int i = 0; i < figures[playerID].Count; i++)
+        {
+            if (figures[playerID][i].GetComponent<FigureParameter>().GetPosition() == pcNodeID[playerID][1])
+            {
+
+                backBenchFigure = figures[playerID][i];
+            }
+            else if (figures[playerID][i].GetComponent<FigureParameter>().GetPosition() == pcNodeID[playerID][0])
+            {
+                moveAnotherPCFigure = figures[playerID][i];
+            }
+        }
+
+        //PC1からベンチへ移動
+        if (backBenchFigure != null)
+        {
+            yield return backBenchFigure.GetComponent<FigureController>().FigureOneStepWalk(backBenchFigure.GetComponent<FigureParameter>().GetBenchID());
+        }
+        //PC0からPC1へ移動
+        if (moveAnotherPCFigure != null)
+        {
+            yield return moveAnotherPCFigure.GetComponent<FigureController>().FigureOneStepWalk(pcNodeID[playerID][1]);
+        }
+        //フィールドからPC0へ移動
+        yield return _figure.GetComponent<FigureController>().FigureOneStepWalk(pcNodeID[playerID][0]);
+    }
+
+    void SceneLoaded(Scene nextScene, LoadSceneMode mode)
+    {
+        if (nextScene.name == "SpinScene")
+        {
+            //ノードの色を初期化
+            for (int i = 0; i < 28; i++)
+            {
+                nodes.transform.GetChild(i).GetComponent<SpriteRenderer>().color = Color.white;
+            }
+            //ノードを非表示
+            for (int i = 0; i < 44; i++)
+            {
+                nodes.transform.GetChild(i).GetComponent<SpriteRenderer>().enabled = false;
+            }
+            //エッジを非表示
+            for (int i = 0; i < this.transform.childCount; i++)
+            {
+                transform.GetChild(i).GetComponent<LineRenderer>().enabled = false;
+            }
+
+            //フィギュアを非表示にする
+            for (int i = 0; i < 2; i++)
+            {
+                foreach (GameObject obj in figures[i])
+                {
+                    for (int j = 0; j < obj.transform.childCount; j++)
+                    {
+                        obj.transform.GetChild(j).GetComponent<SpriteRenderer>().enabled = false;
+                    }
+                }
+            }
+
+            //UIの非表示
+            turnEndButton.SetActive(false);
+            //UIカット
+            //restTurnText.GetComponent<TextMeshProUGUI>().enabled = false;
+
+        }
+        //SpinSceneからBoardSceneに帰ってきた
+        else if (nextScene.name == "BoardScene")
+        {
+            Destroy(GameObject.Find("BeforeStart"));
+            //ノードを再表示
+            for (int i = 0; i < 44; i++)
+            {
+                nodes.transform.GetChild(i).GetComponent<SpriteRenderer>().enabled = true;
+            }
+            //エッジを再表示
+            for (int i = 0; i < this.transform.childCount; i++)
+            {
+                transform.GetChild(i).GetComponent<LineRenderer>().enabled = true;
+            }
+            //フィギュアを再表示
+            for (int i = 0; i < 2; i++)
+            {
+                foreach (GameObject obj in figures[i])
+                {
+                    for (int j = 0; j < obj.transform.childCount; j++)
+                    {
+                        obj.transform.GetChild(j).GetComponent<SpriteRenderer>().enabled = true;
+                    }
+                }
+            }
+            //UIの表示
+            turnEndButton.SetActive(true);
+            //UIカット
+            //restTurnText.GetComponent<TextMeshProUGUI>().enabled = true;
+
+            
+            //バトル後の処理
+            StartCoroutine(SetPhaseState(PhaseState.AfterBattle));
+        }
+    }
+
+    IEnumerator FadeInOut(GameObject _text, float _fadetime)
+    {
+        float a = 0f;
+        while (a < _fadetime)
+        {
+            _text.GetComponent<TextMeshProUGUI>().color += new Color(0, 0, 0, Time.deltaTime);
+            a += Time.deltaTime;
+            yield return null;
+        }
+        while (a > 0)
+        {
+            _text.GetComponent<TextMeshProUGUI>().color -= new Color(0, 0, 0, Time.deltaTime);
+            a -= Time.deltaTime;
+            yield return null;
+        }
+
+    }
+
+    public void turnEnd()
+    {
+        StartCoroutine(SetPhaseState(PhaseState.TurnEnd));
+    }
 }
