+++ Assets/Assets/Scripts/BoardMaster/BoardController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2fc5f84a199a5f44f99c386cdcbe46b5
+guid: 019e7521d953a95458d7eae432ef006b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
