+++ Assets/Assets/Scripts/Figure/FigureController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3a693b78c561bf04497f3f6f9db27829
+guid: b8a719ccc2aee1e48be8881181e48985
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
