+++ Assets/Assets/Scripts/Nodes/NodeController.cs
@@ -21,6 +21,7 @@ public class NodeController : MonoBehaviour
     
     public void OnUserAction()
     {
+        Debug.Log(nodeParameter.GetNodeID());
         StartCoroutine(boardController.NodeClicked(nodeParameter.GetNodeID()));
         //boardController.NodeClicked(nodeParameter.GetNodeID());
     }
