+++ Assets/Assets/Scripts/Nodes/NodeController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2a5d926b4320526479adb6867a1afa49
+guid: c546b5a2ced76964c96a874801183f3b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
