+++ Assets/Assets/Scripts/Other/BeforeGameBlind.cs
