+++ Assets/Assets/Scripts/SpinMaster/SpinController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 687d885ff7cab694e9c9f308dda80d5f
+guid: ca398c3b2a005a042ab7fafb54a4df25
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
