--- Assets/Assets/Scripts/_BoardScene2D.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 907fa4ff5dd9d2a4da71c8c3b87745e8
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
