--- Assets/Assets/Scripts/_BoardScene2D/BoardController2D.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 019e7521d953a95458d7eae432ef006b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
