--- Assets/Assets/Scripts/_BoardScene2D/FigureController2D.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b8a719ccc2aee1e48be8881181e48985
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 200
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
