--- Assets/Assets/Scripts/_BoardScene2D/NodeController2D.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c546b5a2ced76964c96a874801183f3b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
