--- Assets/Assets/Scripts/_BoardScene2D/SpinController2D.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ca398c3b2a005a042ab7fafb54a4df25
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
