--- Assets/Assets/Test.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4f49e3a45221e46458d26044ef77d246
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
