+++ ProjectSettings/EditorBuildSettings.asset
@@ -27,9 +27,9 @@ EditorBuildSettings:
     path: Assets/Assets/Scenes/TitleScene_2.unity
     guid: a24850eeda57c0d4db96f09d21acf8a5
   - enabled: 1
-    path: Assets/Assets/Scenes/BoardScene2D.unity
+    path: Assets/Assets/Scenes/BoardScene.unity
     guid: 6baec6ffff48adb4faffcf5dd7e0750b
   - enabled: 1
-    path: Assets/Assets/Scenes/SpinScene2D.unity
+    path: Assets/Assets/Scenes/SpinScene.unity
     guid: 34eed99dfb865414285f004eaec9073e
   m_configObjects: {}
